+++ ydb/core/tx/schemeshard/schemeshard_scheme_builders.h
@@ -6,10 +6,6 @@ namespace NKikimrScheme {
     class TEvDescribeSchemeResult;
 }
 
-namespace NKikimrSchemeOp {
-    enum EPathType : int;
-}
-
 namespace NKikimr::NSchemeShard {
 
 bool BuildScheme(
