+++ ydb/core/tx/schemeshard/schemeshard_xxport__helpers.h
@@ -7,6 +7,14 @@ namespace Ydb::Operations {
     class OperationParams;
 }
 
+namespace NKikimr::NBackup {
+    enum class EBackupFileType : unsigned char;
+}
+
+namespace NKikimrSchemeOp {
+    enum EPathType : int;
+}
+
 namespace NKikimr::NSchemeShard {
 
 class TSchemeShard;
@@ -21,5 +29,15 @@ TString GetUid(const Ydb::Operations::OperationParams& operationParams);
 
 THolder<TEvSchemeShard::TEvModifySchemeTransaction> MakeModifySchemeTransaction(TSchemeShard* ss, TTxId txId, const TExportInfo& exportInfo);
 THolder<TEvSchemeShard::TEvModifySchemeTransaction> MakeModifySchemeTransaction(TSchemeShard* ss, TTxId txId, const TImportInfo& importInfo);
+
+struct XxportProperties {
+    TString FileName;
+    NBackup::EBackupFileType FileType;
+    NKikimrSchemeOp::EPathType PathType;
+};
+
+const TVector<XxportProperties>& GetXxportProperties();
+
+TMaybe<XxportProperties> PathTypeToXxportProperties(NKikimrSchemeOp::EPathType pathType);
 
 }  // NKikimr::NSchemeShard
